Restructure tree and clean up codebase
