+++ plugin.video.vrt.nu/lib/vrtplayer/metadatacollector.py
@@ -2,7 +2,7 @@
 
 # GNU General Public License v2.0 (see COPYING or https://www.gnu.org/licenses/gpl-2.0.txt)
 
-from resources.lib.vrtplayer import metadatacreator
+from lib.vrtplayer import metadatacreator
 
 
 class MetadataCollector:
