+++ plugin.video.vrt.nu/lib/vrtplayer/metadatacreator.py
