+++ plugin.video.vrt.nu/lib/vrtplayer/statichelper.py
