+++ plugin.video.vrt.nu/lib/vrtplayer/tokenresolver.py
@@ -2,12 +2,12 @@
 
 # GNU General Public License v2.0 (see COPYING or https://www.gnu.org/licenses/gpl-2.0.txt)
 
-from resources.lib.helperobjects import helperobjects
-import requests
-import json
 import datetime
+import json
+import requests
 import time
 
+from lib.helperobjects import helperobjects
 
 class TokenResolver:
