+++ test/streamservicetests.py
@@ -2,9 +2,10 @@
 
 # GNU General Public License v2.0 (see COPYING or https://www.gnu.org/licenses/gpl-2.0.txt)
 
-import unittest
-from resources.lib.vrtplayer import vrtplayer, streamservice, tokenresolver
 from mock import MagicMock
+import unittest
+
+from lib.vrtplayer import streamservice, tokenresolver, vrtplayer
 
 use_drm = False
